+++ .idea/encodings.xml
@@ -13,6 +13,8 @@
     <file url="file://$PROJECT_DIR$/fe-blog-system-gateway/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/fe-blog-system-index/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/fe-blog-system-index/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/fe-blog-system-media/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/fe-blog-system-media/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/fe-blog-system-ornament/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/fe-blog-system-ornament/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/fe-blog-system-user/src/main/java" charset="UTF-8" />
