+++ fe-blog-system-bean/src/main/java/com/bloducspauter/bean/blog/BlogTag.java
@@ -1,13 +1,11 @@
-package com.bloducspauter.bean;
+package com.bloducspauter.bean.blog;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
-import com.mysql.cj.xdevapi.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 
 import java.io.Serializable;
 
@@ -16,8 +14,8 @@ import java.io.Serializable;
 @AllArgsConstructor
 @Data
 @TableName
-public class Tag implements Serializable {
+public class BlogTag implements Serializable {
     @TableId(type = IdType.AUTO)
+    private Integer blogId;
     private Integer tagId;
-    private String name;
 }
