+++ fe-blog-system-bean/src/main/java/com/bloducspauter/bean/blog/Homepage.java
@@ -1,4 +1,4 @@
-package com.bloducspauter.bean;
+package com.bloducspauter.bean.blog;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
