+++ fe-blog-system-bean/src/main/java/com/bloducspauter/bean/category/Field.java
@@ -1,7 +1,8 @@
-package com.bloducspauter.bean;
+package com.bloducspauter.bean.category;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,8 +13,9 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-public class MediaRelation implements Serializable {
+@TableName
+public class Field implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Integer blogId;
-    private Integer mediaId;
+    private Integer fieldId;
+    private String name;
 }
