+++ fe-blog-system-bean/src/main/java/com/bloducspauter/bean/category/TagRelation.java
@@ -1,4 +1,4 @@
-package com.bloducspauter.bean;
+package com.bloducspauter.bean.category;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
