+++ fe-blog-system-bean/src/main/java/com/bloducspauter/bean/media/MediaRelation.java
@@ -1,8 +1,7 @@
-package com.bloducspauter.bean;
+package com.bloducspauter.bean.media;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,10 +12,8 @@ import java.io.Serializable;
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
-@TableName
-public class Media implements Serializable {
+public class MediaRelation implements Serializable {
     @TableId(type = IdType.AUTO)
+    private Integer blogId;
     private Integer mediaId;
-    private String image;
-    private String music;
 }
