+++ fe-blog-system-blog/src/main/java/com/bloducspauter/blog/mapper/HomePageMapper.java
@@ -2,7 +2,7 @@ package com.bloducspauter.blog.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
-import com.bloducspauter.bean.Homepage;
+import com.bloducspauter.bean.blog.Homepage;
 
 @Mapper
 public interface HomePageMapper extends BaseMapper<Homepage> {
