+++ fe-blog-system-blog/src/main/java/com/bloducspauter/blog/service/impl/BlogServiceImpl.java
@@ -2,12 +2,16 @@ package com.bloducspauter.blog.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 
-import com.bloducspauter.bean.*;
+import com.bloducspauter.bean.blog.Blog;
+import com.bloducspauter.bean.blog.BlogTag;
+import com.bloducspauter.bean.category.Tag;
+import com.bloducspauter.bean.category.TagRelation;
+import com.bloducspauter.bean.media.Media;
 import com.bloducspauter.blog.mapper.BlogMapper;
 import com.bloducspauter.blog.service.BlogService;
 
 
-import com.bloducspauter.category.mapper.MediaMapper;
+import com.bloducspauter.ornament.mapper.MediaMapper;
 import com.bloducspauter.category.mapper.TagMapper;
 import com.bloducspauter.category.mapper.TagRelationMapper;
 import lombok.extern.slf4j.Slf4j;
