+++ fe-blog-system-blog/src/main/java/com/bloducspauter/blog/service/impl/HomePageServiceImpl.java
@@ -1,7 +1,7 @@
 package com.bloducspauter.blog.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.bloducspauter.bean.Homepage;
+import com.bloducspauter.bean.blog.Homepage;
 import com.bloducspauter.blog.mapper.HomePageMapper;
 import com.bloducspauter.blog.service.HomePageService;
 import lombok.extern.slf4j.Slf4j;
