+++ fe-blog-system-blog/src/main/resources/bootstrap.yml
@@ -3,7 +3,7 @@ spring:
     nacos:
       config:
 #        server-addr: 172.18.0.2:8848
-        server-addr: 192.168.24.128:8848
+        server-addr: 192.168.43.49:8848
         namespace: 192099489
         group: FE_BLOG
         prefix: fe-blog-blog
@@ -17,18 +17,4 @@ spring:
       discovery:
         namespace: 192099489
         group: FE_BLOG
-        server-addr: 192.168.24.128:8848
-
-
-#server:
-#  port: 51809
-#spring:
-#  application:
-#    name: redis-local
-#  cloud:
-#    nacos:
-#      discovery:
-#        server-addr: 192.168.24.128:8848
-#      config:
-#        server-addr: 192.168.24.128:8848
-#        file-extension: yml
+        server-addr: 192.168.43.49:8848
