+++ fe-blog-system-category/src/main/java/com/bloducspauter/category/mapper/FieldMapper.java
@@ -2,12 +2,8 @@ package com.bloducspauter.category.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 
-import com.bloducspauter.bean.Blog;
-import com.bloducspauter.bean.Field;
+import com.bloducspauter.bean.category.Field;
 import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.util.List;
 
 
 @Mapper
