+++ fe-blog-system-category/src/main/java/com/bloducspauter/category/mapper/TagRelationMapper.java
@@ -2,7 +2,7 @@ package com.bloducspauter.category.mapper;
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 
-import com.bloducspauter.bean.TagRelation;
+import com.bloducspauter.bean.category.TagRelation;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
