+++ fe-blog-system-category/src/main/resources/bootstrap.yml
@@ -3,7 +3,7 @@ spring:
     nacos:
       config:
 #        server-addr: 172.18.0.3:8848
-        server-addr: 192.168.24.128:8848
+        server-addr: 192.168.43.49:8848
         namespace: 192099489
         username: nacos
         password: nacos
@@ -13,35 +13,6 @@ spring:
           - {data-id: 'mysql8-local.yml',refresh: true,group: FE_BLOG}
           - {data-id: 'common.yml',refresh: true,group: FE_BLOG}
           - {data-id: 'redis-local',refresh: false,group: FE_BLOG}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
         group: FE_BLOG
       discovery:
        namespace: 192099489
