+++ fe-blog-system-gateway/src/main/resources/bootstrap.yml
@@ -3,7 +3,7 @@ spring:
     nacos:
       config:
 #        server-addr: 172.18.0.2:8848
-        server-addr:  http://192.168.24.128:8848
+        server-addr:  http://192.168.43.49:8848
         namespace: 192099489
         username: nacos
         password: nacos
@@ -16,4 +16,4 @@ spring:
         namespace: 192099489
         group: FE_BLOG
         service: fe-gateway
-        server-addr:  http://192.168.24.128:8848
+        server-addr:  http://192.168.43.49:8848
