+++ fe-blog-system-media/pom.xml
@@ -20,4 +20,15 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-autoconfigure</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.bloducspauter</groupId>
+            <artifactId>fe-blog-system-bean</artifactId>
+            <version>1.0</version>
+        </dependency>
+    </dependencies>
 </project>
