+++ fe-blog-system-media/src/main/java/com/bloducspauter/media/MediaApplication.java
@@ -1,0 +1,16 @@
+package com.bloducspauter.media;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Media启动类
+ * @author Bloduc Spauter
+ *
+ */
+@SpringBootApplication(scanBasePackages = {"com.bloducspauter.media","com.bloducspauter.base"})
+public class MediaApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(MediaApplication.class, args);
+    }
+}
