+++ fe-blog-system-ornament/pom.xml
@@ -112,6 +112,11 @@
             <artifactId>sigar</artifactId>
             <version>1.6.4</version>
         </dependency>
+        <dependency>
+            <groupId>com.bloducspauter</groupId>
+            <artifactId>fe-blog-system-bean</artifactId>
+            <version>1.0</version>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
