+++ fe-blog-system-ornament/src/main/java/com/bloducspauter/ornament/FeBlogOrnamentApp.java
@@ -1,12 +1,15 @@
-package com.bloducspauter.category;
+package com.bloducspauter.ornament;
 
+import com.bloducspauter.ornament.service.MediaService;
+import com.bloducspauter.user.service.UploadService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = {"com.bloducspauter.ornament", "com.bloducspauter.bean"},
+        scanBasePackageClasses= {UploadService.class, MediaService.class})
 @EnableDiscoveryClient
 @EnableFeignClients(basePackages = {"com.bloducspauter.api"})
 @EnableRedisHttpSession
