+++ fe-blog-system-ornament/src/main/java/com/bloducspauter/ornament/controller/GetInfoController.java
@@ -1,4 +1,4 @@
-package com.bloducspauter.category.controller;
+package com.bloducspauter.ornament.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.hyperic.sigar.SigarException;
