+++ fe-blog-system-ornament/src/main/java/com/bloducspauter/ornament/mapper/MediaMapper.java
@@ -1,8 +1,8 @@
-package com.bloducspauter.category.mapper;
+package com.bloducspauter.ornament.mapper;
 
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.bloducspauter.bean.Media;
+import com.bloducspauter.bean.media.Media;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
