+++ fe-blog-system-ornament/src/main/resources/mapper/CommentMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.bloducspauter.category.mapper.CommentMapper">
+<mapper namespace="com.bloducspauter.ornament.mapper.CommentMapper">
 
 
     <select id="selectAllComment" resultType="com.bloducspauter.bean.Comment">
