+++ fe-blog-system-ornament/src/main/resources/mapper/ReplyMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.bloducspauter.category.mapper.ReplyMapper">
+<mapper namespace="com.bloducspauter.ornament.mapper.ReplyMapper">
     <select id="selectReplyByPage" resultType="com.bloducspauter.bean.Reply">
         select * from reply where id = #{id} limit #{page},#{size}
     </select>
