+++ fe-blog-system-user/src/main/resources/bootstrap.yml
@@ -3,7 +3,7 @@ spring:
     nacos:
       config:
 #        server-addr: 172.18.0.2:8848
-        server-addr: http://192.168.24.128:8848
+        server-addr: http://192.168.43.49:8848
         namespace: 192099489
         username: nacos
         password: nacos
@@ -18,4 +18,4 @@ spring:
         group: FE_BLOG
         namespace: 192099489
         service: fe-user
-        server-addr:  http://192.168.24.128:8848
+        server-addr:  http://192.168.43.49:8848
